Fix line split in secure mount and notify
Currently the code splits the mount command line returns
incorrectly. This results in a non match which means a new
mount is created everytime, instead of using an existing mount.

This results in many mounts being present on the system, every
time the agent is restarted.

This is due to the split element being a back slash `\` whereas
it should be a whitespace `' '`

There is also no feedback provided to the user when an existing
mount is used. This information should be provided to the user.

Resolves: #83