Generate Cargo.lock file
The Cargo FAQ explain that to determine deterministic builds of binary
files, is recommended to commit the Cargo.lock files in binary projects.

This becomes relevant when building a RPM package for distribution, when
all the depenencies are generated via "cargo vendor".  The build will
fail if "Cargo.lock" is not present.

This commit removes the Cargo.lock file from .gitignore and generate a
Cargo.lock file using the Rust 1.53.0 release.

Signed-off-by: Alberto Planas <aplanas@suse.com>