ima: Read IMA measurement list begining at n-th entry.
Implement read_measurement_list that allows us to read the IMA measurement
log from a certain entry number. Since the last log entry *may not* be part
of the quote the validator may ask for the last entry in the log again,
otherwise it will try to read the next entry following what it already had
seen (n+1 entry after seeing n entries). Therefore, we have to keep track
of the log's number of entries and corresponding filesizes at the last query
and the previous to last query so that we can use seek() to easily find the
appropriate position in the file without having to walk or send the whole
measurement log again.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>