move global configuration into a single struct
The KeylimeConfig struct is introduced to hold the configuration data that
is currently used by the agent.

All config_get* functions are now private and should not be used outside of
the common module.

If possible functions outside of the main module should not take the
KeylimeConfig as an argument. Instead only pass the necessary data as
separate arguments. Type conversions should be done by the caller.

For testing purposes the Default trait is implemented for KeylimeConfig.
Use this instead of loading the configuration from disk.

Signed-off-by: Thore Sommer <mail@thson.de>