revocation: Use revocation certificate set by configuration (#300)
Obtain the certificate path according to the value set to the
revocation_cert entry in the configuration.  When the value is
"default", the default path is used.  When the value is a relative path,
the path is expanded from the WORK_DIR.  When the value is an absolute
path, use it as provided.  In case the value is empty, return error.

This follows the python agent implementation behavior.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>