Set working directory via KEYLIME_DIR env variable
The working directory can be set via the KEYLIME_DIR environment
variable or keylime_dir configuration option.

When the working directory is set, the expected location for the
certificates and secure mount are also modified and are relative to the
location set via KEYLIME_DIR.

The DEFAULT_CA_PATH is now relative to the working directory.

Consolidate the conditional change of the secure mount location when
testing to a single location.

Modify the revocation tests to use a temporary directory as the working
directory.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>