keys_handler: Add API to verify derived key
Add the handler to the agent for GET requests to '/keys/verify'
endpoint.

The handler expects the 'challenge' parameter as 20 alphanumeric
characters string and responds with a JSON structure containing the
resulting hex-encoded HMAC-SHA384 calculated using the symmetric key
derived from U and V keys and the challenge as the input.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>