Use JSON wrapper on error responses
The response for requests should always be wrapped in the JSON
structure, including in error cases.

Without this, the rust agent returns malformed responses on error cases,
advertising `content-type: application/json` on the header, but sending
plaintext on the body, which caused JSON parsing errors.

This was found during implementation of
https://github.com/keylime/keylime/pull/947.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>