main: Add default handlers for each scope
The added handlers are called when a request do not match with previous
services in each scope.  This may happen if an unsupported method or an
unsupported URI is used.

When an unsupported method is invoked, a 405 error is returned and the
ALLOW header is set in the response with the supported methods as
specified in [RFC 7231], section 6.5.5.

This also adds a handler for unsupported API versions.

[RFC 7231](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.5)

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>