Make wiremock an optional dependency
Make a wiremock optional, enabled when the 'testing' feature is enabled.
The goal is to reduce the number of dependencies, with the downside that
the test code coverage will drop.

Wiremock is currently used only on 'registrar_agent' tests to mock
responses from the registrar service.

Note: wiremock was moved to be an optional regular dependency because
optional dev-dependencies are not supported.
see: https://github.com/rust-lang/cargo/issues/1596

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>