Update clap dependency to 3.1.18
This change replaces clap::App with clap::Command in main.rs file.

On commit f3fa925 the clap dependency in Cargo.lock file was updated
from version 3.0.14 to version 3.1.18.

In clap version 3.1.0, clap::App was deprecated in favor of
clap::Command.  This causes warnings during the Keylime agent build due
to the usage of the deprecated structure.

To avoid similar issues in the future, use the tilde dependency version
requirement to allow versions >= 3.1.18 but < 3.2.0.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>