tpm: Add serialization functions for structures in quotes
Add serialization and deserialization functions for the data in quotes
to avoid endianness issues when the arch is big-endian.

The added serialization and deserialization functions will convert the
data endianness as necessary.

Note: the official marshalling and unmarshalling functions cannot be
used directly because the tpm2-tools uses a custom format.

Fixes: #407

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>