Allow using existing key and certificate
Separate the key and certificate from the Agent Data and store in
separate files.

Add the configuration options 'server_key' and 'server_cert' to allow
the user to use existing key and certificate.

If the files set to 'server_key' or 'server_cert' don't exist, new files
are created to store the generated key or certificate.  Otherwise the
existing files will not be overwritten.

This also unify the handling of file paths in the configuration file
for the files stored in the 'keylime_dir' by introducing the
config_get_file_path() function with the following behavior:

- If the option is set as "default" or left empty, the default value is
  used
- If the option is not an absolute path, it is treated as relative from
  the directory set in the 'keylime_dir' configuration option
- If the option is set with an absolute path, it is used without change

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>