Add support for using passphrase protected key
Add the configuration option server_key_password to allow setting a
password to be used to encrypt the generated key when writing to a file
and to decrypt the key when loading from a file.

If the server_key_password is set as empty, the key is assumed to be
unencrypted.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>