Use password to generate EK when provided
When tpm_ownerpassword option is set in the configuration file, set the
password in the TPM context for the Endorsement Hierarchy regardless of
ek_handle value set in the configuration file.

This way, when the tpm_ownerpassword option is set in the configuration
file, it is used when accessing the TPM to generate the Endorsement
primary key handle.

The password in the TPM context for the Owner Hierarchy is not set
anymore as we do not use it.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>