Add build script to set `feature=testing` cfg
This is necessary when running the keylime/keylime CI with run_local.sh,
which indicates a test with the env var KEYLIME_TEST set to 'true'
(as a string). Otherwise, the MOUNT_SECURE flag in common.rs is set
incorrectly when running this with the CI as opposed to in a Rust-based
testing environment.

Signed-off-by: Lily Sturmann <lsturman@redhat.com>