main: only read uuid from KeylimeConfig
The main function also had a local variable called uuid. This variable
contained either the provided uuid or when "hash_ek" was set the sha256 hash
of the EK. We never updated the KeylimeConfig struct which caused a mismatch
between uuids used in the program. Instead we now mutate the KeylimeConfig
when "hash_ek" is set to the correct value.

Signed-off-by: Thore Sommer <mail@thson.de>