Refactor async code for keys and payloads
Use messages for synchronization instead of mutexes and condvar.  Add
separate workers for keys and payload execution, holding the state
locally instead of storing all data in server context (QuoteData).

The keys worker holds the received U and K keys and is responsible for
combining them to obtain K.  It also provides the K key when requested
by the /keys/verify endpoint to prove that the resulting K is correct.

When the U and V keys combination is successful (the HMAC verification
is successful), the keys worker wakes the payloads worker sending the
obtained K key and the encrypted payload.

The payloads worker is responsible for decrypting and running the
payload sent by the keys worker.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>