config: Refactor AgentConfig Source trait implementation
Instead of using explicit dedicated transformations for each option in
the configuration structure, use an intermediary JSON structure to
implement the collect() method in a more generic way.

This makes it easier to add new configuration options and simplifies
maintenance.

Also, instead of duplicating the AgentConfig structure to receive the
options set in the environment variables, use the built-in Environment
Source trait implementation with a small change to make it possible to
use as a source for a KeylimeConfig structure.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>