tests: Add script to generate IAK and IDevID certificates
The script generates a test root CA and intermediate CA keys and
certificates. The output directory can be set via the --output option,
and the CA password with the --pwd option.  If an output directory is
not set, a temporary directory is created and the certificates are
placed there.

Then, the script generates the IDevID and IAK keys inside the TPM and
respective certificates.

The script requires the tpm2-openssl provider to access the keys inside
the TPM.

It is recommended to set the TCTI, TPM2TOOLS_TCTI, and TPM2OPENSSL_TCTI
environment variables to set which TPM to use when running the script.
If these variables are not set, the default value `device:/dev/tpmrm0`
is used instead.

This also modifies the tests/setup_swtpm.sh and tests/run.sh to
configure the TPM2OPENSSL_TCTI to use the swtpm with the tpm2-openssl
provider.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>