config: Make IAK and IDevID certificates optional
When IAK/IDevID are enabled, but the paths to the certificates are
explicitly configured as the empty string, continue normally and
register without IAK and IDevID certificates.

This is to make it possible to use IAK and IDevID without the
certificates, in case the user does the public key matching check
separately.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>