main: Support using multiple API versions for registration
Use the keylime::registrar_client module instead of the registrar_agent,
which is deleted.

This enables the agent to communicate with a registrar using an older
API version, restoring the backwards compatibility.

This also removes the unnecessary `API_VERSION` from `common.rs`.

This is part of the implementation of the enhancement proposal 114:
https://github.com/keylime/enhancements/pull/115

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>