version: Implement API version validation and ordering
Validate the values set in the `api_versions` configuration option, and
filter only the supported versions.

The configured versions are also sorted so that the agent can try the
enabled versions from the newest to the oldest.

If none of the configured options are supported, fallback to use all the
supported API versions instead.

This is part of the implementation of the enhancement proposal 114:
https://github.com/keylime/enhancements/pull/115

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>