lib/cert: Fix race condition due to use of same file path
The tests could enter in a race condition where one test removes the
file while the other is using it. This was caused by the files using the
same path.

This uses a temporary directory for each test.  This makes it
unnecessary to manually remove the files and also avoids the risk of
race conditions since each temporary directory has a unique name.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>