config: Allow using alternative keylime_dir for tests
Some tests assume the `/var/lib/keylime` will always be present, which
is not the case if Keylime is not installed.

Add the `get_testing_config()` method to get a configuration where the
keylime_dir can be set to be another path.

The new method is only compiled when the feature `testing` is enabled.
This means that tests that want to use this method should be restricted
to run when the `testing` feature is enabled.

The new method should replace the calls to KeylimeConfig::default() in
the tests to avoid failing during the validation that the path set in
`keylime_dir` option, by default `/var/lib/keylime`, exists.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>