tests: Do not rely on the existence of /var/lib/keylime
Many tests would assume that the `/var/lib/keylime` directory exists,
which may not be the case when Keylime is not installed.

Calls to KeylimeConfig::default() would result in failure because the
`config_translate_keywords()` method would verify if the directory set
in the `keylime_dir` option exists (by default `/var/lib/keylime`).

This replaces the KeylimeConfig::default() calling with
`get_testing_config()` in tests to set an existing temporary directory
as the `keylime_dir`.

Alternatively, when the `KeylimeConfig` structure is constructed
manually, the `keylime_dir` option is set as an existing temporary
directory.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>