tests: Do not assume `/var/lib/keylime` exists (#1030)
* config: Allow using alternative keylime_dir for tests

Some tests assume the `/var/lib/keylime` will always be present, which
is not the case if Keylime is not installed.

Add the `get_testing_config()` method to get a configuration where the
keylime_dir can be set to be another path.

The new method is only compiled when the feature `testing` is enabled.
This means that tests that want to use this method should be restricted
to run when the `testing` feature is enabled.

The new method should replace the calls to KeylimeConfig::default() in
the tests to avoid failing during the validation that the path set in
`keylime_dir` option, by default `/var/lib/keylime`, exists.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>

* tests: Do not rely on the existence of /var/lib/keylime

Many tests would assume that the `/var/lib/keylime` directory exists,
which may not be the case when Keylime is not installed.

Calls to KeylimeConfig::default() would result in failure because the
`config_translate_keywords()` method would verify if the directory set
in the `keylime_dir` option exists (by default `/var/lib/keylime`).

This replaces the KeylimeConfig::default() calling with
`get_testing_config()` in tests to set an existing temporary directory
as the `keylime_dir`.

Alternatively, when the `KeylimeConfig` structure is constructed
manually, the `keylime_dir` option is set as an existing temporary
directory.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>

---------

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>