config: Remove KeylimeConfig structure
Remove KeylimeConfig structure, making AgentConfig the only
configuration structure.  The KeylimeConfig structure was just a wrapper
around AgentConfig.

This also moves the loading of configuration files to the file_config.rs
file.

A test for overriding options using environment variables is added as a
integration test. This is necessary to make it to run in a separate
process and not affect other tests.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>