context_info_handler: Do not assume /var/lib/keylime exists
This changes the tests in the context_info_handler.rs to not assume the
/var/lib/keylime exists.

In tests, whenever a configuration is necessary, the
`config::get_testing_config()` should be called, passing the path to a
temporary directory as argument so that it overrides the `keylime_dir`
option to not use the default `/var/lib/keylime`.

Note that this is only available when the `testing` feature is enabled.

This is necessary to avoid the tests to fail when the `/var/lib/keylime`
directory is not present, for example in build systems or when the tests
are executed before Keylime is installed.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>