keylime/structures: Rename ShaValues to PcrBanks
Add all the supported hash algorithms, including the missing sha384,
sha512, and sm3_256 to the PcrBanks structure.

Implement the builder pattern to create the structure more
ergonomically, which will not set empty banks to the structure.

Skip the serialization of unset banks, avoiding empty banks to be sent
to the Verifier.

When populating the requests in struct_filler.rs, use the supported
algorithms obtained from the TPM and populate the PcrBanks structure
using the builder.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>