Add 6 alphanumeric lowercase X-Request-ID header
This PR enhances the ResilientClient by automatically adding a unique
X-Request-ID header to every outgoing HTTP request. This is a common de
facto standard that significantly improves request tracing, correlation,
and debugging across distributed systems.

Key Changes

* Unique ID Generation:
  - A new private function, generate_request_id, has been added to
create a random, 6-character, lowercase alphanumeric string to serve as
a unique identifier.
  - The rand crate was added as a dependency to support this
functionality.

* Automatic Header Injection:
  - Both get_request and get_json_request methods have been modified
to call generate_request_id and attach the resulting ID to the request
via the X-Request-ID header.

* Mockoon tests:
  - Mockoon has been upgraded with a specific route (x-request-id-test)
so that it only responds 200 OK if the header exists and contains
correct format

* Logging:
  - Sent request URL, Method and headers are logged in debug mode

Signed-off-by: Sergio Arroutbi <sarroutb@redhat.com>