Use simple algorithm specifiers in certification_keys object (#1140)
Update key_algorithm reporting in the certification_keys object to use
simple specifiers ("rsa" or "ecc") instead of full algorithm names
("rsa2048", "ecc256", etc.) to comply with the push model protocol
specification.

This is not an issue for algorithm identification because the
certification_keys object also includes the key_size field, allowing
complete identification of the exact algorithm in use.

Signed-off-by: Sergio Correia <scorreia@redhat.com>