Persist payload keys to avoid attestation failure on restart
Previously, when the agent was restarted while being monitored by the
verifier, it would fail attestation because the PCR#16 would be extended
with the new ephemeral payload public key.

With this change, the payload key is persisted, making the agent to
reuse the previously generated payload key instead of generating a new
ephemeral key on restart.

Resolves: #1136

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>